+++ VevaciousPlusPlus/include/PotentialEvaluation/ThermalFunctions.hpp
@@ -132,7 +132,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( -10.0 * ( squareRatio + 1.0 ) );
     size_t floorIndex( scaledRatio );
     return ( bosonMinusOneToMinusTwelve[ floorIndex ]
-             + ( 0.1 * ( scaledRatio - (double)floorIndex )
+             + ( 0.1 * ( scaledRatio - floorIndex )
                      * ( bosonMinusOneToMinusTwelve[ floorIndex + 1 ]
                          - bosonMinusOneToMinusTwelve[ floorIndex ] ) ) );
   }
@@ -144,7 +144,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( -100.0 * squareRatio );
     size_t floorIndex( scaledRatio );
     return ( bosonZeroToMinusOne[ floorIndex ]
-             + ( 0.01 * ( scaledRatio - (double)floorIndex )
+             + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( bosonZeroToMinusOne[ floorIndex + 1 ]
                           - bosonZeroToMinusOne[ floorIndex ] ) ) );
   }
@@ -156,7 +156,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( 100.0 * squareRatio );
     size_t floorIndex( scaledRatio );
     return ( bosonZeroToPlusOne[ floorIndex ]
-             + ( 0.01 * ( scaledRatio - (double)floorIndex )
+             + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( bosonZeroToPlusOne[ floorIndex + 1 ]
                           - bosonZeroToPlusOne[ floorIndex ] ) ) );
   }
@@ -168,7 +168,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( squareRatio - 1.0 );
     size_t floorIndex( scaledRatio );
     return ( bosonPlusOneToPlusOneHundred[ floorIndex ]
-             + ( ( scaledRatio - (double)floorIndex )
+             + ( ( scaledRatio - floorIndex )
                  * ( bosonPlusOneToPlusOneHundred[ floorIndex + 1 ]
                      - bosonPlusOneToPlusOneHundred[ floorIndex ] ) ) );
   }
@@ -180,7 +180,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( -10.0 * ( squareRatio + 1.0 ) );
     size_t floorIndex( scaledRatio );
     return ( fermionMinusOneToMinusTwelve[ floorIndex ]
-             + ( 0.1 * ( scaledRatio - (double)floorIndex )
+             + ( 0.1 * ( scaledRatio - floorIndex )
                      * ( fermionMinusOneToMinusTwelve[ floorIndex + 1 ]
                          - fermionMinusOneToMinusTwelve[ floorIndex ] ) ) );
   }
@@ -192,7 +192,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( -100.0 * squareRatio );
     size_t floorIndex( scaledRatio );
     return ( fermionZeroToMinusOne[ floorIndex ]
-             + ( 0.01 * ( scaledRatio - (double)floorIndex )
+             + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( fermionZeroToMinusOne[ floorIndex + 1 ]
                           - fermionZeroToMinusOne[ floorIndex ] ) ) );
   }
@@ -204,7 +204,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( 100.0 * squareRatio );
     size_t floorIndex( scaledRatio );
     return ( fermionZeroToPlusOne[ floorIndex ]
-             + ( 0.01 * ( scaledRatio - (double)floorIndex )
+             + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( fermionZeroToPlusOne[ floorIndex + 1 ]
                           - fermionZeroToPlusOne[ floorIndex ] ) ) );
   }
@@ -216,7 +216,7 @@ namespace VevaciousPlusPlus
     double scaledRatio( squareRatio - 1.0 );
     size_t floorIndex( scaledRatio );
     return ( fermionPlusOneToPlusOneHundred[ floorIndex ]
-             + ( ( scaledRatio - (double)floorIndex )
+             + ( ( scaledRatio - floorIndex )
                  * ( fermionPlusOneToPlusOneHundred[ floorIndex + 1 ]
                      - fermionPlusOneToPlusOneHundred[ floorIndex ] ) ) );
   }
